move `apply` into `TypeCheck.External`
