+++ lib/type_check.ex
@@ -322,79 +322,4 @@ defmodule TypeCheck do
       {:error, exception} -> raise exception
     end
   end
-
-  @doc """
-  The same as `Kernel.apply/3` but ensures that values conform the given type.
-
-  The first argument must be a function type of the function to be called.
-
-  In case of type error, raises an exception.
-  
-  ## Examples
-
-      iex> alias TypeCheck.Builtin, as: B
-      iex> type = B.function([B.number], B.number)
-      iex> TypeCheck.apply!(type, Kernel, :abs, [-13])
-      13
-      iex> TypeCheck.apply!(type, Kernel, :abs, ["hello"])
-      ** (TypeCheck.TypeError) At lib/type_check.ex:279:
-          `"hello"` is not a number.
-
-  """
-  @spec apply!(TypeCheck.Type.t(), module(), atom(), list()) :: any()
-  def apply!(type, module, function, args) do
-    {ptypes, rtype} = split_func_type(type)
-    with :ok <- check_params(args, ptypes),
-         result <- Kernel.apply(module, function, args),
-         _ <- dynamic_conforms!(result, rtype) do
-      result
-    else
-      {:error, reason} -> raise reason
-    end
-  end
-
-  @doc """
-  The same as `TypeCheck.apply/3` but ensures that values conform the given type.
-
-  Returns `{:error, reason}` on failure, `{:ok, function_call_result}` otherwise.
-  
-      iex> alias TypeCheck.Builtin, as: B
-      iex> type = B.function([B.number], B.number)
-      iex> TypeCheck.apply(type, Kernel, :abs, [-13])
-      {:ok, 13}
-      iex> {:error, err} = TypeCheck.apply(type, Kernel, :abs, [false])
-      iex> err.message
-      "At lib/type_check.ex:279:\\n    `false` is not a number."      
-
-  """
-  @spec apply(TypeCheck.Type.t(), module(), atom(), list()) ::
-          {:ok, any()} | {:error, TypeCheck.TypeError.t()}
-  def apply(type, module, function, args) do
-    {ptypes, rtype} = split_func_type(type)
-    with :ok <- check_params(args, ptypes),
-         result <- Kernel.apply(module, function, args),
-         :ok <- dynamic_conforms(result, rtype) do
-      result
-    end
-  end
-
-  @spec check_params(list(), [TypeCheck.Type.t()]) :: :ok | {:error, TypeCheck.TypeError.t()}
-  defp check_params([value | values], [type | types]) do
-    case dynamic_conforms(value, type) do
-      {:ok, _} -> check_params(values, types)
-      {:error, reason} -> {:error, reason}
-    end
-  end
-
-  defp check_params(_, _), do: :ok
-  
-  @spec split_func_type(TypeCheck.Type.t()) :: {[TypeCheck.Type.t()], TypeCheck.Type.t()}
-  defp split_func_type(%{param_types: ptypes, return_type: rtype}), do: {ptypes, rtype}
-
-  defp split_func_type(%{__struct__: TypeCheck.Builtin.OneOf, choices: types}) do
-    ptypes = types |> Enum.map(fn %{param_types: params} -> params end) |> List.zip
-      |> Enum.map(fn types -> types |> Tuple.to_list |> TypeCheck.Builtin.one_of end)
-    rtypes = Enum.map(types, fn %{return_type: rtype} -> rtype end)
-    {TypeCheck.Builtin.one_of(ptypes), TypeCheck.Builtin.one_of(rtypes)}
-  end
 end
